Fixed some issues with the generation

+++ src/main.rs
@@ -147,7 +147,7 @@ impl App {
                 match game_state.input(args, &mut self.context) {
                     ControlFlow::Break(MainMenuOption::NewGame) => {
                         self.scene = SceneEnum::WorldGen(WorldGenScene::new(WorldGenerationParameters {
-                            seed: Rng::rand().rand_u32(),
+                            seed: Rng::rand().rand_u32() as u64,
                             world_size: Size2D(64, 48),
                             history_length: 2000,
                             number_of_seed_cities: 3,
